% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{captionize}
\alias{captionize}
\title{captionize}
\usage{
captionize(str, split = " |_|\\\\.")
}
\arguments{
\item{str}{a string or a vector of strings}

\item{split}{regex to split the string. Defaults to ' |_|\\.'}
}
\description{
This function splits the given str with the split regex and concatenate them after capitalizing every word
}
\examples{
captionize('where_is_waldo')
}
\keyword{capital}
\keyword{capitalize}
\keyword{caption}
