% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colorset64}
\alias{colorset64}
\title{Color Set 64}
\format{A vector of RGB colors:}
\usage{
colorset64
}
\description{
This is a set of 64 completely distingushable colors for plotting
}
\keyword{datasets}
