% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{createColoredBarGraphLabels}
\alias{createColoredBarGraphLabels}
\title{splitCol}
\usage{
createColoredBarGraphLabels(data, col.x, col.fill, col.facet = NULL,
  middle = TRUE, reverse = TRUE)
}
\arguments{
\item{data}{a data frame one of its columns to be split}

\item{col.x}{the column name used on x aes}

\item{col.fill}{the column name used for fill/color aes}

\item{col.facet}{the column name used for facet aes}

\item{middle}{create the pos to be in the middle. defaults TRUE}

\item{reverse}{create the pos on reverse. defaults TRUE}
}
\description{
This function splits the given column in the data.frame
}
\examples{
infert$c <- factor(infert$induced)
texts <-  createColoredBarGraphLabels(infert, 'education', 'c')
p <- ggplot(infert, aes(education, fill=c)) + geom_bar(alpha=.9)
p     
p + geom_text(data=texts, aes(y=pos, label=cnt))    
p + geom_text(data=texts, aes(y=pos, label=goodNum(100*cnt/tot, percent=TRUE)))    
}
\keyword{barplot}
\keyword{ggplot}
