% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{extract_field}
\alias{extract_field}
\title{extract_field}
\usage{
extract_field(data, regex)
}
\arguments{
\item{data}{a vector of char objects}

\item{regex}{character containing regular expression(s) with a group in it}
}
\description{
This function extracts the first group of regex from a vector
}
\examples{
extract_field('hello_23_node', '(\\\\d+)')   # '23'
extract_field(c('all-2', 'any-3', 'never-4-d'), '(\\\\d+)') # a vector c('2', '3', '4')
extract_field(c('all-2', 'any+3', 'never+4'), '\\\\+(\\\\d+)') # a vector c(NA, '3', '4')
}
\keyword{expression}
\keyword{extract}
\keyword{extract_field}
\keyword{regular}
