% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{goodNum}
\alias{goodNum}
\title{goodNum}
\usage{
goodNum(numbers, digits = 2, percent = FALSE, ...)
}
\arguments{
\item{numbers}{a number (float or int) or a vector of numbers}

\item{digits}{number of digits after fraction for float numbers. Defaults to 2.}

\item{percent}{(bool) show the % sign at the end of the number. Defaults to FALSE.}
}
\description{
This function creates a nice representation of a given (vector of) number.
}
\examples{
goodNum(c(12, 22.1, 112112.119))
goodNum(c(12000, 1002000, 3454))
}
\keyword{formatC}
\keyword{formating}
