% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{if.na}
\alias{if.na}
\title{if.na}
\usage{
if.na(data, subs = "")
}
\arguments{
\item{data}{an object to be checked if it's NA}

\item{subs}{substitute the NA values with subs (Defauts empty string)}
}
\description{
This function sreplaces the NA in a list with a value
}
\examples{
if.na(airquality$Ozone, -1)
}
\keyword{if.na}
\keyword{ifna}
