% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{if.null}
\alias{if.null}
\title{if.null}
\usage{
if.null(data, subs = "")
}
\arguments{
\item{data}{an object to be checked if it's NULL}

\item{subs}{substitute the NULL values with subs (Defauts empty string)}
}
\description{
This function sreplaces the NULL in a list with a value
}
\examples{
if.null(airquality$Ozone, -1)
}
\keyword{if.null}
\keyword{ifnull}
\keyword{null}
