% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{splitCol}
\alias{splitCol}
\title{splitCol}
\usage{
splitCol(data, col_name, sep, fixed = FALSE)
}
\arguments{
\item{data}{a data frame one of its columns to be split}

\item{col_name}{the column whose data to be split}

\item{sep}{regex to split the data.}

\item{fixed}{regex to split the data.}
}
\description{
This function splits the given column in the data.frame
}
\examples{
splitCol(infert, 'education', '-')
}
\keyword{column}
\keyword{split}
\keyword{splitCol}
