% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{splitText}
\alias{splitText}
\title{splitText}
\usage{
splitText(vec, sep, part = 0, fixed = FALSE, fill = TRUE)
}
\arguments{
\item{vec}{a vector of char objects}

\item{sep}{character containing regular expression(s) (unless fixed = TRUE) to use for splitting.}

\item{part}{the part of the split to return, when 0 returns all. defaults to 0.}

\item{fixed}{logical. If TRUE match split exactly, otherwise use regular expressions. defaults to FALSE.}

\item{fill}{logical. fill with NA when needed.  defaults to TRUE.}
}
\description{
This function splits the given column in the data.frame
}
\examples{
splitText('hello2all4of222you', '\\\\d+')   # a vector ('hello','all', 'of', 'you')
splitText('hello.all.of.you', '.', fixed=TRUE)  # same as above
splitText('hello-all-of-you', '-')   # same as above
splitText('hello-all-of-you', '-', 2) # 'all'
splitText(c('all-2', 'any-3', 'never-4'), '-') # a data.frame with 2 columns
splitText(c('all-2', 'any-3', 'never-4-d'), '-') # The third column will be NA for the first two rows
splitText(c('all-2', 'any-3', 'never-4'), '-', 1) # a vector c('all', 'any', 'never')
}
\keyword{column}
\keyword{split}
\keyword{splitCol}
